004 study completed (1D dynamic blend value change through custom node to node connection)
This study show that you can use animation kernel node to connect and pass information to another kernel node.

I've create a BlendNode which retrieve the PlayerBlendKernelRuntime blend value and feed it to the MixerNode Weight value in the  KernelPort.

Animation Kernel utilize the burst compiler and use Jobs so it's better to use kernel then directly setting the data in a Node because of it not being able to use jobs and the burst compiler.

created UI using UIDocument and hook up a slider event when slider changed to component which is then fed to the Blend Node.