006 study complete 2D blend using BlendTree 8 way animation + idle
finished study on how to use a 2D BlendTree using the animation. It is very similar to the 1D BlendTree study, except the you must create a blob asset reference to a blend2DSimple directional and the switch the Blend1DNode to Blend2DNode and pass  2 input parameter to the Blend2DNode kernelport to specify the weight on the x axis and the y axis. Other then that it is the same.